refactor: (팀 컨벤션 변경) 가독성 위해 커스텀훅 케밥케이스에서 카멜케이스로 변경
