+++ apps/web/src/features/auth/signup/lib/useEmailDuplication.ts
@@ -3,7 +3,7 @@ import type { Control, UseFormClearErrors, UseFormSetError } from 'react-hook-fo
 import { useEmailAvailabilityQuery } from '../api';
 import { FORM_MESSAGES } from '../config/form-messages';
 import type { SignupFormValues } from '../model/signup-schema';
-import { useDuplicationCheck } from './use-duplication-check';
+import { useDuplicationCheck } from './useDuplicationCheck';
 
 export function useEmailDuplication(
   control: Control<SignupFormValues>,
