+++ apps/web/src/pages/moments-new/index.ts
@@ -1,0 +1,1 @@
+export { MomentsNewPage } from './ui/MomentsNewPage';
