+++ apps/web/src/pages/moments-new/ui/MomentsNewPage.tsx
@@ -1,0 +1,3 @@
+export function MomentsNewPage() {
+  return <div>Moments New Page</div>;
+}
