+++ apps/web/src/shared/model/loading/index.ts
@@ -1,3 +1,3 @@
-export { GlobalLoadingController } from './GlobalLoadingController';
+export { GlobalLoadingController } from './global-loading-controller';
 export { useGlobalLoadingState } from './useGlobalLoadingState';
 export { useRouteLoading } from './useRouteLoading';
