+++ apps/web/src/shared/model/loading/useGlobalLoadingState.ts
@@ -2,7 +2,7 @@
  * 전역 로딩 노출 상태 조회
  * 로딩 상태에 반응해야 하는 컴포넌트에서 사용
  */
-import { useLoadingStore } from './GlobalLoadingController';
+import { useLoadingStore } from './global-loading-controller';
 
 export function useGlobalLoadingState(): { isVisible: boolean } {
   const isVisible = useLoadingStore((state) => state.isVisible);
