+++ apps/web/src/shared/ui/link-card/LinkCard.tsx
