+++ apps/web/src/widgets/layout/authenticated-shell/ui/index.tsx
@@ -8,7 +8,7 @@ import {
   usePushPermissionSheet,
 } from '@/src/features/push-notifications';
 import { isIosUserAgent, isMobileUserAgent } from '@/src/shared/lib/device/user-agent';
-import { usePwaInstallState } from '@/src/shared/lib/pwa/use-pwa-install-state';
+import { usePwaInstallState } from '@/src/shared/lib/pwa/usePwaInstallState';
 import { ROUTE_GROUPS, ROUTES } from '@/src/shared/routes';
 import { PwaInstallBottomSheet } from '@/src/widgets/pwa-install';
